+++ app/src/main/java/com/magnet/messagingsample/activities/UserSelectActivity.java
@@ -51,59 +51,17 @@ public class UserSelectActivity extends AppCompatActivity {
 
     UsersRecyclerViewAdapter adapter;
 
-    private AtomicBoolean mLoginSuccess = new AtomicBoolean(false);
-    private CallbackManager mCallbackManager = null;
-    private MyProfile mProfile = null;
-    private FacebookCallback<LoginResult> mLoginCallback = new FacebookCallback<LoginResult>() {
-        public void onSuccess(LoginResult loginResult) {
-            Log.d(TAG, "LoginCallback.onSuccess() " + loginResult.getAccessToken().getUserId());
-            final GraphRequest request = new GraphRequest(loginResult.getAccessToken(), "me");
-            Bundle params = new Bundle();
-            params.putString("fields", "id,name,email");
-            request.setParameters(params);
-            AsyncTask.execute(new Runnable() {
-                public void run() {
-                    Log.d(TAG, "LoginCallback.onSuccess() lookup email");
-                    GraphResponse response = request.executeAndWait();
-                    JSONObject jsonResp = response.getJSONObject();
-                    try {
-                        String email = jsonResp.getString("email");
-                        String name = jsonResp.getString("name");
-                        String id = jsonResp.getString("id");
-                        mProfile.setUsername(email);
-                        mProfile.setPassword(id.getBytes()); //FIXME: Find a better way to generate this password
-                        mProfile.setDisplayName(name);
-                        setupMMX();
-                    } catch (final Exception ex) {
-                        Log.e(TAG, "LoginCallback.onSuccess(): exception caught while getting identity", ex);
-                        runOnUiThread(new Runnable() {
-                            public void run() {
-                                Toast.makeText(UserSelectActivity.this, "Exception caught while getting identity: " +
-                                        ex.getMessage(), Toast.LENGTH_LONG).show();
-                                UserSelectActivity.this.finish();
-                            }
-                        });
-                    }
-                }
-            });
-        }
-
-        public void onCancel() {
-            Log.d(TAG, "LoginCallback.onCancel()");
-            UserSelectActivity.this.finish();
-        }
-
-        public void onError(FacebookException e) {
-            Log.e(TAG, "LoginCallback.onError() ", e);
-            Toast.makeText(UserSelectActivity.this, "Exception: " + e.getMessage(), Toast.LENGTH_LONG).show();
-            UserSelectActivity.this.finish();
-        }
-    };
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_user_select);
+
+        if (Profile.getCurrentProfile() == null) {
+            MMX.logout(null);
+            Intent intent;
+            intent = new Intent(UserSelectActivity.this, LoginActivity.class);
+            startActivity(intent);
+        }
 
         rvUsers = (RecyclerView) findViewById(R.id.rvUsers);
 
@@ -114,19 +72,8 @@ public class UserSelectActivity extends AppCompatActivity {
         layoutManager.setOrientation(LinearLayoutManager.VERTICAL);
         rvUsers.setLayoutManager(layoutManager);
 
-        mProfile = MyProfile.getInstance(this);
-        FacebookSdk.sdkInitialize(getApplicationContext());
-        mCallbackManager = CallbackManager.Factory.create();
-        LoginManager.getInstance().registerCallback(mCallbackManager, mLoginCallback);
-        if (mProfile.getUsername() == null || Profile.getCurrentProfile() == null) {
-            //go login
-            HashSet<String> permissions = new HashSet<String>();
-            permissions.add("user_friends");
-            permissions.add("email");
-            LoginManager.getInstance().logInWithReadPermissions(this, permissions);
-        } else {
-            setupMMX();
-        }
+        updateViewState();
+
         // TODO: this doesnt return all users
 //        HashSet names = new HashSet();
 //        names.add("");
@@ -146,7 +93,6 @@ public class UserSelectActivity extends AppCompatActivity {
 //                refreshListView(null);
 //            }
 //        });
-
     }
 
     private void updateViewState() {
@@ -159,47 +105,6 @@ public class UserSelectActivity extends AppCompatActivity {
             public void onFailure(MMXUser.FailureCode failureCode, Throwable throwable) {
                 Log.e(TAG, "MMXUser.findByName() error: " + failureCode, throwable);
                 refreshListView(null);
-            }
-        });
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        mCallbackManager.onActivityResult(requestCode, resultCode, data);
-    }
-
-    private void setupMMX() {
-        // Register this activity as a listener to receive and show incoming
-        // messages.  See #onDestroy for the unregister call.
-        MMXUser internalUser = new MMXUser.Builder()
-                .username(mProfile.getUsername())
-                .displayName(mProfile.getDisplayName())
-                .build();
-        internalUser.register(mProfile.getPassword(), new MMXUser.OnFinishedListener<Void>() {
-            public void onSuccess(Void aVoid) {
-                Log.d(TAG, "register user succeeded");
-                loginHelper();
-            }
-
-            public void onFailure(MMXUser.FailureCode failureCode, Throwable throwable) {
-                Log.d(TAG, "register user failed because: " + failureCode);
-                loginHelper();
-            }
-        });
-    }
-
-    private void loginHelper() {
-        MMX.login(mProfile.getUsername(), mProfile.getPassword(), new MMX.OnFinishedListener<Void>() {
-            public void onSuccess(Void aVoid) {
-                mLoginSuccess.set(true);
-                MMX.enableIncomingMessages(true);
-                updateViewState();
-            }
-
-            public void onFailure(MMX.FailureCode failureCode, Throwable e) {
-                mLoginSuccess.set(false);
-                updateViewState();
             }
         });
     }
