+++ app/src/main/res/layout/activity_login.xml
@@ -42,7 +42,7 @@
         android:textColor="#000000"
         android:layout_below="@+id/ivLogo"
         android:layout_centerHorizontal="true"
-        android:layout_marginTop="45dp" />
+        android:layout_marginTop="30dp" />
 
     <EditText
         android:layout_width="wrap_content"
@@ -82,6 +82,20 @@
         android:layout_alignRight="@+id/etPassword"
         android:layout_alignEnd="@+id/etPassword" />
 
+    <com.facebook.login.widget.LoginButton
+        android:id="@+id/btFacebookLogin"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:layout_below="@+id/btLogin"
+        android:layout_marginTop="15dp"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true"
+        android:paddingTop="15dp"
+        android:paddingBottom="15dp" />
+
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -89,7 +103,7 @@
         android:id="@+id/btRegister"
         android:background="#03A9F4"
         android:textColor="#fff"
-        android:layout_below="@+id/btLogin"
+        android:layout_below="@+id/btFacebookLogin"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
         android:layout_alignParentRight="true"
