+++ app/src/main/res/values/strings.xml
@@ -6,6 +6,7 @@
 
     <string name="title_activity_login">Rich Messaging</string>
     <string name="login_btn_text">Log In</string>
+    <string name="login_facebook_btn_text">Login With Facebook</string>
     <string name="register_btn_text">Register</string>
     <string name="email_label_text">Username</string>
     <string name="password_text_label">Password</string>
@@ -16,4 +17,5 @@
     <string name="title_activity_image_view">Image</string>
     <string name="title_activity_video_view">Video</string>
 
+    <string name="facebook_app_id">894485373938499</string>
 </resources>
