--- app/src/main/res/values/strings_facebook.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="facebook_app_id">894485373938499</string>
-</resources>
